sudoku_flask/requirements.txt
absl-py==0.10.0
asgiref @ file:///tmp/build/80754af9/asgiref_1594338739818/work
astunparse==1.6.3
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cycler==0.10.0
decorator==4.4.2
Flask==1.1.2
gast==0.3.3
google-auth==1.21.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.31.0
h5py==2.10.0
idna==2.10
imageio==2.9.0
importlib-metadata==1.7.0
imutils==0.5.3
itsdangerous==1.1.0
Jinja2==2.11.2
Keras==2.4.3
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.3.1
mkl-fft==1.1.0
mkl-random==1.1.1
mkl-service==2.3.0
networkx==2.5
numpy @ file:///C:/ci/numpy_and_numpy_base_1596233948451/work
oauthlib==3.1.0
opt-einsum==3.3.0
Pillow==7.2.0
protobuf==3.13.0
py-sudoku==1.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
PyWavelets==1.1.1
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
scikit-image==0.17.2
scipy @ file:///C:/ci/scipy_1597675683670/work
six==1.15.0
sqlparse==0.3.1
sudoku-manager==1.0.6
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.0
tensorflow-estimator==2.3.0
termcolor==1.1.0
tifffile==2020.9.3
urllib3==1.25.10
Werkzeug==1.0.1
wincertstore==0.2
wrapt==1.12.1
zipp==3.1.0
